% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2dna.R
\name{string2dna}
\alias{string2dna}
\title{String to DNA encoder/decoder}
\usage{
string2dna(x, method = "encode")
}
\arguments{
\item{x}{A string}

\item{method}{Either \code{encode} (turn a string into a nucleotide sequence), or \code{decode} (turn a nucleotide sequence into a string).}
}
\value{
A string
}
\description{
Encode a string into a DNA sequence or decode a DNA sequence into a string.
}
\examples{
string2dna("Stephen Turner", method="encode")
string2dna("TGATAGCTAGAGCGTCTACTCAGCTAGGATCACCTCCTACAC", method="decode")
}
